--- esmvaltool/diag_scripts/autoassess/autoassess_source/__init__.py
@@ -1,1 +1,0 @@
-"""Initialize for autoassess_source"""
